+++ appnongsan/lib/screens/account.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:appnongsan/screens/mainscreen.dart';
+import 'package:appnongsan/screens/demoscreen.dart';
 import 'question.dart';
 import 'info.dart';
